// types/AuthType.ts
export enum AuthType {
  None = 0,
  UAS_ID_Signature = 1,
  Operator_ID_Signature = 2,
  Message_Set_Signature = 3,
  Network_Remote_ID = 4,
  Specific_Authentication = 5,
  Private_Use_0xA = 0xa,
  Private_Use_0xB = 0xb,
  Private_Use_0xC = 0xc,
  Private_Use_0xD = 0xd,
  Private_Use_0xE = 0xe,
  Private_Use_0xF = 0xf,
}